+++ src/lib/components/editor/index.ts
@@ -1,3 +1,4 @@
 export { default as TrackEditor } from './TrackEditor.svelte'
 export { default as EditorField } from './EditorField.svelte'
+export { default as EditorTextArea } from './EditorTextArea.svelte'
 export { default as EditorArtwork } from './EditorArtwork.svelte'